requirements.txt
requests
beautifulsoup4
urllib3
colorama
tqdm
html2text
rich
playwright
PyYAML
